% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to the database using SSH}
\usage{
connect()
}
\value{
list of length 2 with r background process and DBI connection object
}
\description{
Connect to the database using SSH
}
\details{
Prerequisites for successfully calling this function:
\enumerate{
\item Store DB username and password in operating system credential store
using the keyring package:
keyring::key_set(service = "ecometri_caddis,
username = "myusername")
\item Generate an ssh key
}
}
