% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_data_tools.R
\name{correctForSpikes}
\alias{correctForSpikes}
\title{Correct isotope data for unlabeled nitrate spikes}
\usage{
correctForSpikes(isotope_data, nitrate_data, spike_table, C_sp, AP_sp)
}
\arguments{
\item{isotope_data}{data table of atom percents to be corrected}

\item{nitrate_data}{data table of nitrate concentrations for samples in isotope_data}

\item{spike_table}{table with sample IDs of spiked samples, volume of sample, and volume of spike}

\item{C_sp}{Nitrate concentration of spike solution}

\item{AP_sp}{Atom percent (N15-Nitrate/(N14-Nitrate + N15 Nitrate)) of spike solutions}
}
\value{

}
\description{
Correct isotope data for unlabeled nitrate spikes
}
