% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_data_tools.R
\name{create_batches}
\alias{create_batches}
\title{Create batches of lab data in the data base}
\usage{
create_batches(cn, labDataList)
}
\arguments{
\item{cn}{MariaDB connection object}

\item{labDataList}{list of lists returned by read_labData}
}
\value{

}
\description{
Create batches of lab data in the data base
}
