% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_data_tools.R
\name{load_labData}
\alias{load_labData}
\title{Load data from lab into the database}
\usage{
load_labData(cn, labData)
}
\arguments{
\item{cn}{MariaDB connection object}

\item{labData}{list of lab data as returned by read_labData}
}
\value{

}
\description{
Load data from lab into the database
}
