% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_data_tools.R
\name{load_loggerData}
\alias{load_loggerData}
\title{Upload conductivity data to database}
\usage{
load_loggerData(cn, data)
}
\arguments{
\item{cn}{DBI connection object}

\item{data}{tibble of readings,as produced by e.g. read_HOBODir or read_miniDOTDir}
}
\value{
inserts records into the database and returns the number of records inserted
}
\description{
Upload conductivity data to database
}
