% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{loadLoggerData}
\alias{loadLoggerData}
\title{Upload conductivity data to database}
\usage{
loadLoggerData(cn, data)
}
\arguments{
\item{cn}{DBI connection object}

\item{data}{tibble of readings,as produced by e.g. read_HOBODir or read_miniDOTDir}
}
\value{
inserts records into the database and returns the number of records inserted
}
\description{
Upload conductivity data to database
}
