% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{loadLoggerDataList}
\alias{loadLoggerDataList}
\title{Load a list of tibbles into the database}
\usage{
loadLoggerDataList(dataList)
}
\arguments{
\item{dataList}{list of tibbles}
}
\value{
inserts records into the database and returns a tibble with
the number of records inserted for each data chunk
}
\description{
Load a list of tibbles into the database
}
