% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{read_HOBOCondData}
\alias{read_HOBOCondData}
\title{Read in a single HOBO conductivity logger text file}
\usage{
read_HOBOCondData(serialNo, fileName, dirName)
}
\arguments{
\item{serialNo}{Logger serial number}

\item{fileName}{name of file}

\item{dirName}{name of directory containing file}
}
\value{
tibble with columns for datetime, temperature, conductivity, and serial number
}
\description{
Read in a single HOBO conductivity logger text file
}
