% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_data_tools.R
\name{read_HOBODir}
\alias{read_HOBODir}
\title{Read in and combine multiple HOBO conductivity logger
text files within a single directory}
\usage{
read_HOBODir(startDate, endDate, loggers, dirName, datatype = "Conductivity")
}
\arguments{
\item{startDate}{earliest date when the logger was programmed with format \%m/\%d/\%y}

\item{endDate}{latest date when the logger was programmed with format \%m/\%d/\%y}

\item{loggers}{Vector of serial numbers for which to read in data}

\item{dirName}{name of directory with conductivity text files}

\item{datatype}{one of Conductivity, Light, or Pressure}
}
\value{
tibble with columns for datetime, temperature, conductivity, and serial number
}
\description{
Read in and combine multiple HOBO conductivity logger
text files within a single directory
}
