% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_data_tools.R
\name{read_HOBOLightData}
\alias{read_HOBOLightData}
\title{Read in a single HOBO light logger text file}
\usage{
read_HOBOLightData(serialNo, fileName, dirName)
}
\arguments{
\item{serialNo}{Logger serial number}

\item{fileName}{name of file}

\item{dirName}{name of directory containing file}
}
\value{
tibble with columns for datetime, metric_name, and metric_value
}
\description{
Read in a single HOBO light logger text file
}
