% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_data_tools.R
\name{read_labData}
\alias{read_labData}
\title{Read in a spreadsheet containing data from an analytical lab for a single analyte}
\usage{
read_labData(filepath)
}
\arguments{
\item{filepath}{path to the spreadsheet}
}
\value{
a list with the lab name, analysis date, filename, metric name,
metric units, detection limit, and tibble of lab data.
}
\description{
Read in a spreadsheet containing data from an analytical lab for a single analyte
}
