% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{read_miniDOTData}
\alias{read_miniDOTData}
\title{Read in a single day of MiniDOT data from a single logger}
\usage{
read_miniDOTData(filePath)
}
\arguments{
\item{filePath}{Path of the MiniDOT text file to read in}
}
\value{
Tibble with columns for time (seconds), DO (mg/L),
and temperature (degrees C)
}
\description{
Read in a single day of MiniDOT data from a single logger
}
