% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{read_miniDOTDir}
\alias{read_miniDOTDir}
\title{Read in and concatenate all miniDOT data files for a single logger}
\usage{
read_miniDOTDir(
  dir,
  serialNos,
  startDate = as_date("2021-01-01"),
  endDate = as_date("2022-01-01")
)
}
\arguments{
\item{dir}{path to directory containing all text files
for single logger}

\item{serialNos}{vector of valid serialNumbers}

\item{startDate}{first date of data to read in, as a Date object}

\item{endDate}{last date of data to read in, as a Date object}
}
\value{
Tibble with columns for logger serial number, time (seconds),
DO (mg/L), and temperature (degrees C)
}
\description{
Read in and concatenate all miniDOT data files for a single logger
}
