% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_data_tools.R
\name{split_loggerData}
\alias{split_loggerData}
\title{Break up logger data into a list of tibbles}
\usage{
split_loggerData(data)
}
\arguments{
\item{data}{tibble to split}
}
\value{
A list of tibbles broken up first by date and, if needed,
into smaller subsets with a maximum of 10,000 rows
}
\description{
Break up logger data into a list of tibbles
}
