% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{splitLoggerData}
\alias{splitLoggerData}
\title{Break up logger data by}
\usage{
splitLoggerData(data)
}
\arguments{
\item{data}{}
}
\value{
A list of tibbles broken up first by date and, if needed,
into smaller subsets with a maximum of 10,000 rows
}
\description{
Break up logger data by
}
